pyproject.toml
[build-system]
requires = [
    "setuptools"
]
build-backend = "setuptools.build_meta"


[project]
name = "Pydle"
version = "0.0.1"
requires-python = ">=3.13"
authors = [
    {name = "Anthony Burrow", email = "anthony.r.burrow@gmail.com"},
]
maintainers = [
    {name = "Anthony Burrow", email = "anthony.r.burrow@gmail.com"}
]
description = "Terminal-based idle RPG made in Python."
readme = "README.md"
classifiers=[
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "pynput",
    "prompt_toolkit",
    "pywin32; sys_platform == 'win32'",
    "python-xlib; sys_platform == 'linux'",
    "colorama",
    "platformdirs"
]

[project.urls]
Repository = "https://github.com/anthonyburrow/Pydle.git"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
pydle = "Pydle.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["Pydle", "Pydle.*"]
